Update Withings_Sleep_Sensor.groovy
Add Sleep Sensor and Switch that sets Sleeping on and off.  I use the switch as an easier way to setup a RM rule to trigger my night routines to set the Ring Security when me and my wife is in bed.   The Sleep Sensor I use in my dashboard.